fix(quality): split the VTH workflow-template seed into lookup, graph resolver and orchestration
SeedVthWorkflowTemplates had an overall class complexity of 85 (threshold 50):
one class owned the catalog walk, every OpenRegister read, the result-row shape
coercion and the steps/transitions translation.

- VthSeedLookup owns the three OpenRegister reads (caseType resolution,
  idempotency probe, statusType map) plus the system-principal elevation, with
  one soft-fail query() seam so 'missing service', 'unconfigured schema' and
  'search threw' all collapse to the same empty result in one place.
- VthSeedRowReader owns the result-row coercion (array / {results:[...]} /
  ObjectEntity).
- VthWorkflowGraphResolver owns the steps/transitions translation and the
  deterministic UUID5 ids.
- SeedVthWorkflowTemplates keeps orchestration only.

The pre-existing class-level @SuppressWarnings(PHPMD.CouplingBetweenObjects) is
removed: with the reads behind a seam the class no longer needs it (verified by
measuring with the suppression gone).

Also drops the phpstan ignoreErrors dependency for this code: the raw catalog
step/transition params are now typed array<int, mixed> — which is what
json_decode() actually yields — so the is_array() guards are real checks rather
than PHPDoc-contradicting dead comparisons.

phpmd: SeedVthWorkflowTemplates ExcessiveClassComplexity 85 -> cleared; repo
total 24 -> 23 findings, no new finding anywhere in lib/.
