+++ lib/Repair/SeedVthWorkflowTemplates.php
@@ -17,6 +17,10 @@
  * resolved, the template is logged + skipped (warning only), and the rest
  * of the catalog continues.
  *
+ * This class is orchestration only. The OpenRegister reads live in
+ * {@see \OCA\Procest\Repair\Vth\VthSeedLookup} and the steps/transitions
+ * translation in {@see \OCA\Procest\Repair\Vth\VthWorkflowGraphResolver}.
+ *
  * @category Repair
  * @package  OCA\Procest\Repair
  *
@@ -39,8 +43,9 @@ declare(strict_types=1);
 namespace OCA\Procest\Repair;
 
 use OCA\Procest\AppInfo\Application;
+use OCA\Procest\Repair\Vth\VthSeedLookup;
+use OCA\Procest\Repair\Vth\VthWorkflowGraphResolver;
 use OCA\Procest\Service\SettingsService;
-use OCA\Procest\Service\Support\SearchesObjects;
 use OCA\Procest\Service\WorkflowDefinitionService;
 use OCP\Migration\IOutput;
 use OCP\Migration\IRepairStep;
@@ -49,14 +54,10 @@ use Psr\Log\LoggerInterface;
 /**
  * Repair step that seeds six canonical VTH workflow templates.
  *
- * @SuppressWarnings(PHPMD.CouplingBetweenObjects) — needs OpenRegister + WorkflowDefinitionService.
- *
  * @spec openspec/specs/vth-workflow-templates/spec.md
  */
 class SeedVthWorkflowTemplates implements IRepairStep
 {
-
-    use SearchesObjects;
 
     /**
      * Catalog directory relative to lib/.
@@ -64,16 +65,12 @@ class SeedVthWorkflowTemplates implements IRepairStep
     private const CATALOG_DIR = __DIR__.'/../Settings/seed/vth-workflow-templates';
 
     /**
-     * UUID5 namespace for deterministic step/transition ids derived from
-     * template slug + child slug.
-     */
-    private const NS_UUID = '6ba7b811-9dad-11d1-80b4-00c04fd430c8';
-
-    /**
      * Constructor for SeedVthWorkflowTemplates.
      *
      * @param SettingsService           $settingsService   Settings service for OR access
      * @param WorkflowDefinitionService $definitionService Workflow lifecycle service
+     * @param VthSeedLookup             $lookup            OpenRegister lookups for the seed
+     * @param VthWorkflowGraphResolver  $graphResolver     Steps/transitions resolver
      * @param LoggerInterface           $logger            Logger
      *
      * @return void
@@ -81,6 +78,8 @@ class SeedVthWorkflowTemplates implements IRepairStep
     public function __construct(
         private readonly SettingsService $settingsService,
         private readonly WorkflowDefinitionService $definitionService,
+        private readonly VthSeedLookup $lookup,
+        private readonly VthWorkflowGraphResolver $graphResolver,
         private readonly LoggerInterface $logger,
     ) {
     }//end __construct()
@@ -103,31 +102,15 @@ class SeedVthWorkflowTemplates implements IRepairStep
      * @param IOutput $output The output interface for progress reporting
      *
      * @return void
-
+     *
      * @spec openspec/changes/retrofit-2026-05-24-case-management/tasks.md
      */
     public function run(IOutput $output): void
     {
         $output->info('Seeding VTH workflow templates...');
 
-        if ($this->settingsService->isOpenRegisterAvailable() === false) {
-            $output->warning(
-                'OpenRegister is not available. Skipping VTH workflow templates seed.'
-            );
-            return;
-        }
-
-        if (is_dir(self::CATALOG_DIR) === false) {
-            $output->warning(
-                'VTH workflow templates catalog directory not found at '
-                .self::CATALOG_DIR
-            );
-            return;
-        }
-
-        $files = glob(self::CATALOG_DIR.'/*.json');
-        if ($files === false || $files === []) {
-            $output->warning('No VTH workflow template catalog files found.');
+        $files = $this->catalogFiles(output: $output);
+        if ($files === []) {
             return;
         }
 
@@ -138,48 +121,14 @@ class SeedVthWorkflowTemplates implements IRepairStep
             'failed'    => 0,
         ];
 
-        // This repair step runs without a Nextcloud user session — anonymous
-        // callers are fail-closed by OpenRegister RBAC (#1955) on every
-        // boot. Without this, every caseType/statusType lookup below reads
-        // as empty and every template is (mis)reported as "caseType not
-        // found", never actually seeding. runAsSystem() also covers the
-        // nested WorkflowDefinitionService::createDraft()/publish() calls,
-        // since the elevation is scoped to this callable for the whole
-        // process, not to one ObjectService instance.
-        $objectService = $this->settingsService->getObjectService();
-        $runner        = function () use ($files, &$summary, $output): void {
-            foreach ($files as $file) {
-                try {
-                    $result           = $this->processCatalogFile(file: $file, output: $output);
+        $this->lookup->runElevated(
+            operation: function () use ($files, &$summary, $output): void {
+                foreach ($files as $file) {
+                    $result           = $this->processCatalogFileSafely(file: $file, output: $output);
                     $summary[$result] = ($summary[$result] ?? 0) + 1;
-                } catch (\Throwable $e) {
-                    $summary['failed']++;
-                    $this->logger->error(
-                        'Procest: failed to process VTH workflow template catalog file',
-                        [
-                            'app'       => Application::APP_ID,
-                            'file'      => basename($file),
-                            'exception' => $e->getMessage(),
-                        ]
-                    );
-                    $output->warning(
-                        'Skipping catalog file '.basename($file)
-                        .' due to processing error (see log).'
-                    );
-                }//end try
-            }//end foreach
-        };
-
-        // Seeding runs with system privileges when OpenRegister exposes
-        // runAsSystem(); without an ObjectService there is nothing to elevate,
-        // so the closure is invoked directly.
-        if ($objectService === null) {
-            $runner();
-        }
-
-        if ($objectService !== null) {
-            $this->runAsSystemIfAvailable(objectService: $objectService, operation: $runner);
-        }
+                }
+            }
+        );
 
         $output->info(
             'VTH workflow templates seed complete: '
@@ -189,6 +138,75 @@ class SeedVthWorkflowTemplates implements IRepairStep
             .$summary['failed'].' failed.'
         );
     }//end run()
+
+    /**
+     * Resolve the catalog files to seed, reporting every precondition that makes
+     * the seed a no-op.
+     *
+     * @param IOutput $output The output interface.
+     *
+     * @return array<int, string> Absolute catalog file paths, or an empty list.
+     *
+     * @spec openspec/specs/vth-workflow-templates/spec.md
+     */
+    private function catalogFiles(IOutput $output): array
+    {
+        if ($this->settingsService->isOpenRegisterAvailable() === false) {
+            $output->warning(
+                'OpenRegister is not available. Skipping VTH workflow templates seed.'
+            );
+            return [];
+        }
+
+        if (is_dir(self::CATALOG_DIR) === false) {
+            $output->warning(
+                'VTH workflow templates catalog directory not found at '
+                .self::CATALOG_DIR
+            );
+            return [];
+        }
+
+        $files = glob(self::CATALOG_DIR.'/*.json');
+        if ($files === false || $files === []) {
+            $output->warning('No VTH workflow template catalog files found.');
+            return [];
+        }
+
+        return $files;
+    }//end catalogFiles()
+
+    /**
+     * Process one catalog file, converting any throw into a `failed` tally.
+     *
+     * One unusable catalog file must never abort the rest of the catalog.
+     *
+     * @param string  $file   Absolute path to the JSON catalog file.
+     * @param IOutput $output The output interface.
+     *
+     * @return string One of seeded|skipped|crossLink|failed
+     *
+     * @spec openspec/specs/vth-workflow-templates/spec.md
+     */
+    private function processCatalogFileSafely(string $file, IOutput $output): string
+    {
+        try {
+            return $this->processCatalogFile(file: $file, output: $output);
+        } catch (\Throwable $e) {
+            $this->logger->error(
+                'Procest: failed to process VTH workflow template catalog file',
+                [
+                    'app'       => Application::APP_ID,
+                    'file'      => basename($file),
+                    'exception' => $e->getMessage(),
+                ]
+            );
+            $output->warning(
+                'Skipping catalog file '.basename($file)
+                .' due to processing error (see log).'
+            );
+            return 'failed';
+        }//end try
+    }//end processCatalogFileSafely()
 
     /**
      * Process a single catalog file.
@@ -212,17 +230,7 @@ class SeedVthWorkflowTemplates implements IRepairStep
         // workflowTemplate; they only document VTH-specific guards that
         // a downstream change should attach to the canonical workflow.
         if ((bool) ($data['crossLink'] ?? false) === true) {
-            $this->logger->info(
-                'Procest: VTH workflow template — cross-link entry, no new workflow created',
-                [
-                    'app'                      => Application::APP_ID,
-                    'slug'                     => $slug,
-                    'targetWorkflowIdentifier' => (string) ($data['targetWorkflowIdentifier'] ?? ''),
-                ]
-            );
-            $output->info(
-                'VTH catalog: cross-link entry "'.$slug.'" — no new workflow created.'
-            );
+            $this->reportCrossLink(data: $data, slug: $slug, output: $output);
             return 'crossLink';
         }
 
@@ -239,7 +247,7 @@ class SeedVthWorkflowTemplates implements IRepairStep
 
         // Resolve steps and transitions. On any unresolved status, skip
         // the entire template (no partial seed).
-        $graph = $this->resolveWorkflowGraph(
+        $graph = $this->graphResolver->resolve(
             data: $data,
             slug: $slug,
             statusMap: (array) $context['statusMap'],
@@ -257,6 +265,30 @@ class SeedVthWorkflowTemplates implements IRepairStep
             output: $output,
         );
     }//end processCatalogFile()
+
+    /**
+     * Report a cross-link catalog entry — documented, never seeded.
+     *
+     * @param array<string, mixed> $data   The decoded catalog entry.
+     * @param string               $slug   The template slug.
+     * @param IOutput              $output The output interface.
+     *
+     * @return void
+     */
+    private function reportCrossLink(array $data, string $slug, IOutput $output): void
+    {
+        $this->logger->info(
+            'Procest: VTH workflow template — cross-link entry, no new workflow created',
+            [
+                'app'                      => Application::APP_ID,
+                'slug'                     => $slug,
+                'targetWorkflowIdentifier' => (string) ($data['targetWorkflowIdentifier'] ?? ''),
+            ]
+        );
+        $output->info(
+            'VTH catalog: cross-link entry "'.$slug.'" — no new workflow created.'
+        );
+    }//end reportCrossLink()
 
     /**
      * Read and validate one catalog file, returning its decoded entry.
@@ -317,38 +349,14 @@ class SeedVthWorkflowTemplates implements IRepairStep
      */
     private function resolveTemplateContext(array $data, string $slug, string $title, IOutput $output): ?array
     {
-        $caseTypeSlug = (string) ($data['caseTypeSlug'] ?? '');
-        if ($caseTypeSlug === '') {
-            $this->logger->warning(
-                'Procest: VTH workflow template — missing caseTypeSlug',
-                ['app' => Application::APP_ID, 'slug' => $slug]
-            );
-            return null;
-        }
-
-        $caseTypeId = $this->resolveCaseTypeId(slug: $caseTypeSlug);
+        $caseTypeId = $this->resolveCaseType(data: $data, slug: $slug, output: $output);
         if ($caseTypeId === '') {
-            // Expected precondition on every boot until base-register-seed-data
-            // has run (or while the anonymous repair context cannot read the
-            // caseType) — debug, not warning, so it does not spam the log.
-            $this->logger->debug(
-                'Procest: VTH workflow template — caseType not found, skipping',
-                [
-                    'app'          => Application::APP_ID,
-                    'slug'         => $slug,
-                    'caseTypeSlug' => $caseTypeSlug,
-                ]
-            );
-            $output->info(
-                'VTH catalog: caseType "'.$caseTypeSlug.'" not found for template "'
-                .$slug.'" — skipping (run base-register-seed-data first).'
-            );
             return null;
         }
 
         // Idempotency: skip if a workflow template with the same title +
         // caseType is already present.
-        if ($this->isAlreadySeeded(caseTypeId: $caseTypeId, title: $title) === true) {
+        if ($this->lookup->isAlreadySeeded(caseTypeId: $caseTypeId, title: $title) === true) {
             $this->logger->info(
                 'Procest: VTH workflow template already present, skipping',
                 [
@@ -361,7 +369,7 @@ class SeedVthWorkflowTemplates implements IRepairStep
         }
 
         // Build the name → UUID map for statusTypes belonging to this caseType.
-        $statusMap = $this->buildStatusMap(caseTypeId: $caseTypeId);
+        $statusMap = $this->lookup->buildStatusMap(caseTypeId: $caseTypeId);
         if ($statusMap === []) {
             $this->logger->warning(
                 'Procest: VTH workflow template — no statusTypes found for caseType',
@@ -381,50 +389,46 @@ class SeedVthWorkflowTemplates implements IRepairStep
     }//end resolveTemplateContext()
 
     /**
-     * Resolve the steps[] and transitions[] blocks against the status map.
+     * Resolve the catalog entry's caseType slug to its UUID.
      *
-     * Returns null when any status name does not resolve — the caller reports
-     * that as skipped (no partial seed).
+     * @param array<string, mixed> $data   The decoded catalog entry.
+     * @param string               $slug   The template slug.
+     * @param IOutput              $output The output interface.
      *
-     * @param array<string, mixed>  $data      The decoded catalog entry
-     * @param string                $slug      The template slug
-     * @param array<string, string> $statusMap Status name → UUID map
-     *
-     * @return array<string, mixed>|null {steps, transitions}, or null when unresolved
+     * @return string The caseType UUID, or the empty string when unresolved.
      */
-    private function resolveWorkflowGraph(array $data, string $slug, array $statusMap): ?array
+    private function resolveCaseType(array $data, string $slug, IOutput $output): string
     {
-        $resolvedSteps = $this->resolveSteps(
-            slug: $slug,
-            rawSteps: ($data['steps'] ?? []),
-            statusMap: $statusMap,
-        );
-        if ($resolvedSteps === null) {
+        $caseTypeSlug = (string) ($data['caseTypeSlug'] ?? '');
+        if ($caseTypeSlug === '') {
             $this->logger->warning(
-                'Procest: VTH workflow template — unresolved status in steps, skipping',
+                'Procest: VTH workflow template — missing caseTypeSlug',
                 ['app' => Application::APP_ID, 'slug' => $slug]
             );
-            return null;
+            return '';
         }
 
-        $resolvedTransitions = $this->resolveTransitions(
-            slug: $slug,
-            rawTransitions: ($data['transitions'] ?? []),
-            statusMap: $statusMap,
-        );
-        if ($resolvedTransitions === null) {
-            $this->logger->warning(
-                'Procest: VTH workflow template — unresolved status in transitions, skipping',
-                ['app' => Application::APP_ID, 'slug' => $slug]
+        $caseTypeId = $this->lookup->resolveCaseTypeId(slug: $caseTypeSlug);
+        if ($caseTypeId === '') {
+            // Expected precondition on every boot until base-register-seed-data
+            // has run (or while the anonymous repair context cannot read the
+            // caseType) — debug, not warning, so it does not spam the log.
+            $this->logger->debug(
+                'Procest: VTH workflow template — caseType not found, skipping',
+                [
+                    'app'          => Application::APP_ID,
+                    'slug'         => $slug,
+                    'caseTypeSlug' => $caseTypeSlug,
+                ]
             );
-            return null;
+            $output->info(
+                'VTH catalog: caseType "'.$caseTypeSlug.'" not found for template "'
+                .$slug.'" — skipping (run base-register-seed-data first).'
+            );
         }
 
-        return [
-            'steps'       => $resolvedSteps,
-            'transitions' => $resolvedTransitions,
-        ];
-    }//end resolveWorkflowGraph()
+        return $caseTypeId;
+    }//end resolveCaseType()
 
     /**
      * Create the draft via the lifecycle service and publish it.
@@ -481,363 +485,4 @@ class SeedVthWorkflowTemplates implements IRepairStep
         $output->info('VTH catalog: seeded "'.$title.'" v'.(int) ($data['version'] ?? 1).'.');
         return 'seeded';
     }//end createAndPublishTemplate()
-
-    /**
-     * Resolve a caseType by its slug — uses the `identifier` field on the
-     * caseType schema (the canonical slug-like field across procest seed
-     * data). Returns the empty string when not found.
-     *
-     * @param string $slug The caseType slug / identifier
-     *
-     * @return string The caseType UUID or empty string
-     */
-    private function resolveCaseTypeId(string $slug): string
-    {
-        $objectService = $this->settingsService->getObjectService();
-        if ($objectService === null) {
-            return '';
-        }
-
-        $register       = $this->settingsService->getConfigValue('register');
-        $caseTypeSchema = $this->settingsService->getConfigValue('case_type_schema');
-
-        if ($register === '' || $caseTypeSchema === '') {
-            return '';
-        }
-
-        // Try `identifier` first (used by bezwaar/beroep seeds), then
-        // `slug` (used by VTH seeds via base-register-seed-data).
-        foreach (['identifier', 'slug'] as $field) {
-            try {
-                $rows = $this->searchObjectsAsArrays(
-                    objectService: $objectService,
-                    register: $register,
-                    schema: $caseTypeSchema,
-                    filters: [$field => $slug, '_limit' => 5],
-                );
-            } catch (\Throwable $e) {
-                $this->logger->debug(
-                    'Procest: VTH workflow template — caseType lookup failed',
-                    [
-                        'app'       => Application::APP_ID,
-                        'field'     => $field,
-                        'slug'      => $slug,
-                        'exception' => $e->getMessage(),
-                    ]
-                );
-                continue;
-            }
-
-            $id = $this->extractFirstId(rows: $rows);
-            if ($id !== '') {
-                return $id;
-            }
-        }//end foreach
-
-        return '';
-    }//end resolveCaseTypeId()
-
-    /**
-     * Check whether a workflowTemplate with the given title is already
-     * present for the given caseType. Used for idempotency.
-     *
-     * @param string $caseTypeId The caseType UUID
-     * @param string $title      The template title
-     *
-     * @return bool True when an existing template matches
-     */
-    private function isAlreadySeeded(string $caseTypeId, string $title): bool
-    {
-        $objectService = $this->settingsService->getObjectService();
-        if ($objectService === null) {
-            return false;
-        }
-
-        $register = $this->settingsService->getConfigValue('register');
-        $schema   = $this->settingsService->getConfigValue('workflow_template_schema');
-
-        if ($register === '' || $schema === '') {
-            return false;
-        }
-
-        try {
-            $rows = $this->searchObjectsAsArrays(
-                objectService: $objectService,
-                register: $register,
-                schema: $schema,
-                filters: [
-                    'caseType' => $caseTypeId,
-                    'title'    => $title,
-                    '_limit'   => 1,
-                ],
-            );
-        } catch (\Throwable $e) {
-            $this->logger->debug(
-                'Procest: VTH workflow template — idempotency lookup failed',
-                [
-                    'app'       => Application::APP_ID,
-                    'caseType'  => $caseTypeId,
-                    'title'     => $title,
-                    'exception' => $e->getMessage(),
-                ]
-            );
-            return false;
-        }//end try
-
-        return $this->extractFirstId(rows: $rows) !== '';
-    }//end isAlreadySeeded()
-
-    /**
-     * Build a status name → UUID map for the statusTypes belonging to a
-     * given caseType.
-     *
-     * @param string $caseTypeId The caseType UUID
-     *
-     * @return array<string, string> Map of statusType name to UUID
-     */
-    private function buildStatusMap(string $caseTypeId): array
-    {
-        $objectService = $this->settingsService->getObjectService();
-        if ($objectService === null) {
-            return [];
-        }
-
-        $register     = $this->settingsService->getConfigValue('register');
-        $statusSchema = $this->settingsService->getConfigValue('status_type_schema');
-
-        if ($register === '' || $statusSchema === '') {
-            return [];
-        }
-
-        try {
-            $rows = $this->searchObjectsAsArrays(
-                objectService: $objectService,
-                register: $register,
-                schema: $statusSchema,
-                filters: ['caseType' => $caseTypeId, '_limit' => 500],
-            );
-        } catch (\Throwable $e) {
-            $this->logger->error(
-                'Procest: VTH workflow template — statusType listing failed',
-                ['app' => Application::APP_ID, 'exception' => $e->getMessage()]
-            );
-            return [];
-        }
-
-        $rowsList = [];
-        if (is_array($rows) === true) {
-            $rowsList = $rows;
-        }
-
-        return $this->mapStatusTypeRows(rows: $rowsList);
-    }//end buildStatusMap()
-
-    /**
-     * Reduce statusType rows to a name → UUID map, dropping unusable rows.
-     *
-     * @param array<int, mixed> $rows The raw statusType rows
-     *
-     * @return array<string, string> Map of statusType name to UUID
-     */
-    private function mapStatusTypeRows(array $rows): array
-    {
-        $map = [];
-        foreach ($rows as $row) {
-            $normalized = $this->normalizeRow(row: $row);
-            if ($normalized === null) {
-                continue;
-            }
-
-            $name = (string) ($normalized['name'] ?? '');
-            $id   = (string) ($normalized['id'] ?? ($normalized['uuid'] ?? ''));
-            if ($name !== '' && $id !== '') {
-                $map[$name] = $id;
-            }
-        }
-
-        return $map;
-    }//end mapStatusTypeRows()
-
-    /**
-     * Resolve the steps[] block against the status name → UUID map.
-     * Returns null when any status name does not resolve.
-     *
-     * @param string                           $slug      The template slug (for UUID5 ids)
-     * @param array<int, array<string, mixed>> $rawSteps  Steps from the catalog file
-     * @param array<string, string>            $statusMap Name → UUID
-     *                                                    map
-     *
-     * @return array<int, array<string, mixed>>|null Resolved steps, or null
-     */
-    private function resolveSteps(string $slug, array $rawSteps, array $statusMap): ?array
-    {
-        $resolved = [];
-        foreach ($rawSteps as $step) {
-            if (is_array($step) === false) {
-                continue;
-            }
-
-            $statusName = (string) ($step['statusName'] ?? '');
-            if ($statusName === '' || isset($statusMap[$statusName]) === false) {
-                return null;
-            }
-
-            $stepSlug   = (string) ($step['slug'] ?? '');
-            $resolved[] = [
-                'id'           => $this->deterministicId(template: $slug, child: 'step-'.$stepSlug),
-                'slug'         => $stepSlug,
-                'title'        => (string) ($step['title'] ?? ''),
-                'status'       => $statusMap[$statusName],
-                'statusName'   => $statusName,
-                'order'        => (int) ($step['order'] ?? 0),
-                'isInitial'    => (bool) ($step['isInitial'] ?? false),
-                'isFinal'      => (bool) ($step['isFinal'] ?? false),
-                'assigneeRole' => ($step['assigneeRole'] ?? null),
-                'description'  => (string) ($step['description'] ?? ''),
-            ];
-        }//end foreach
-
-        return $resolved;
-    }//end resolveSteps()
-
-    /**
-     * Resolve the transitions[] block against the status name → UUID map.
-     * Accepts "*" as a wildcard for fromStatus (any status). Returns null
-     * when any non-wildcard status name does not resolve.
-     *
-     * @param string                           $slug           The template slug (for UUID5 ids)
-     * @param array<int, array<string, mixed>> $rawTransitions Transitions from the catalog file
-     * @param array<string, string>            $statusMap      Name → UUID
-     *                                                         map
-     *
-     * @return array<int, array<string, mixed>>|null Resolved transitions, or null
-     */
-    private function resolveTransitions(string $slug, array $rawTransitions, array $statusMap): ?array
-    {
-        $resolved = [];
-        foreach ($rawTransitions as $transition) {
-            if (is_array($transition) === false) {
-                continue;
-            }
-
-            $fromName = (string) ($transition['fromStatus'] ?? '');
-            $toName   = (string) ($transition['toStatus'] ?? '');
-
-            if ($toName === '' || isset($statusMap[$toName]) === false) {
-                return null;
-            }
-
-            $fromId = '*';
-            if ($fromName !== '*') {
-                if ($fromName === '' || isset($statusMap[$fromName]) === false) {
-                    return null;
-                }
-
-                $fromId = $statusMap[$fromName];
-            }
-
-            $transitionSlug = (string) ($transition['slug'] ?? '');
-            $resolved[]     = [
-                'id'               => $this->deterministicId(template: $slug, child: 'transition-'.$transitionSlug),
-                'slug'             => $transitionSlug,
-                'label'            => (string) ($transition['label'] ?? ''),
-                'fromStatus'       => $fromId,
-                'fromStatusName'   => $fromName,
-                'toStatus'         => $statusMap[$toName],
-                'toStatusName'     => $toName,
-                'allowedRoles'     => ($transition['allowedRoles'] ?? []),
-                'guards'           => ($transition['guards'] ?? []),
-                'automaticActions' => ($transition['automaticActions'] ?? []),
-                'deadline'         => ($transition['deadline'] ?? null),
-            ];
-        }//end foreach
-
-        return $resolved;
-    }//end resolveTransitions()
-
-    /**
-     * Generate a deterministic UUID5 from a template slug + child slug.
-     * Re-running the repair step therefore produces stable step / transition
-     * ids per template.
-     *
-     * @param string $template The template slug
-     * @param string $child    The child slug (e.g. "step-ontvangen")
-     *
-     * @return string The deterministic UUID5
-     */
-    private function deterministicId(string $template, string $child): string
-    {
-        $namespace = str_replace('-', '', self::NS_UUID);
-        $nameBytes = hex2bin($namespace).$template.':'.$child;
-        $hash      = sha1($nameBytes);
-
-        return sprintf(
-            '%08s-%04s-%04x-%04x-%12s',
-            substr($hash, 0, 8),
-            substr($hash, 8, 4),
-            (hexdec(substr($hash, 12, 4)) & 0x0fff) | 0x5000,
-            (hexdec(substr($hash, 16, 4)) & 0x3fff) | 0x8000,
-            substr($hash, 20, 12)
-        );
-    }//end deterministicId()
-
-    /**
-     * Extract the first row id from an OpenRegister result set.
-     *
-     * @param mixed $rows Raw result from searchObjectsAsArrays()
-     *
-     * @return string The first id or empty string
-     */
-    private function extractFirstId(mixed $rows): string
-    {
-        if (is_array($rows) === false) {
-            return '';
-        }
-
-        // Handle paginated `{ results: [...] }` shape.
-        if (isset($rows['results']) === true && is_array($rows['results']) === true) {
-            $rows = $rows['results'];
-        }
-
-        foreach ($rows as $row) {
-            $normalized = $this->normalizeRow(row: $row);
-            if ($normalized === null) {
-                continue;
-            }
-
-            $id = (string) ($normalized['id'] ?? ($normalized['uuid'] ?? ''));
-            if ($id !== '') {
-                return $id;
-            }
-        }
-
-        return '';
-    }//end extractFirstId()
-
-    /**
-     * Coerce an OpenRegister result row to an associative array.
-     *
-     * @param mixed $row Result row from ObjectService
-     *
-     * @return array<string, mixed>|null
-     */
-    private function normalizeRow(mixed $row): ?array
-    {
-        if (is_array($row) === true) {
-            return $row;
-        }
-
-        if (is_object($row) === true && method_exists($row, 'jsonSerialize') === true) {
-            $serialized = $row->jsonSerialize();
-            if (is_array($serialized) === true) {
-                return $serialized;
-            }
-        }
-
-        if (is_object($row) === true && method_exists($row, 'getId') === true) {
-            return ['id' => (string) $row->getId()];
-        }
-
-        return null;
-    }//end normalizeRow()
 }//end class
